refactored package structure to follow MVC
