+++ code/app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:exported="true">
         </activity>
         <activity
-            android:name=".StartupActivity"
+            android:name=".view.StartupActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
