+++ code/app/src/main/java/com/github/bytebandits/bithub/model/MoodPost.java
@@ -1,4 +1,4 @@
-package com.github.bytebandits.bithub;
+package com.github.bytebandits.bithub.model;
 
 import android.graphics.Bitmap;
 import android.location.Location;
